Cleaned up the last merge and lots of changes.
OBS: Some of those features where already on last commit but had not been fully fledged.
- Fixed problems with last merge.
- Remade the tilemaps.
- Added scenes.
- Added audio system.
- Added activator/action scripts.
- Added player death and respawn systems.
- Added pause menu.
- Added lots os sprites and grouped them in files.
- Renamed some scripts to fit the standard.
- Changed main menu to be closer to pause menu.

+++ Treinamento-2018-Grupo1/Assets/Scripts/Utility/SceneScript.cs
@@ -12,12 +12,12 @@ public class SceneScript : MonoBehaviour {
     public struct CameraBoundary {
 
         public bool limitX;
-        public int minX;
-        public int maxX;
+        public float minX;
+        public float maxX;
 
         public bool limitY;
-        public int minY;
-        public int maxY;
+        public float minY;
+        public float maxY;
 
     } public CameraBoundary cameraBoundary; // Limites nos quais a camera dessa cena tem que ficar no eixo-x.
